refactor: moved event types to types/ dir
