+++ src/components/EventCard.astro
@@ -1,5 +1,5 @@
 ---
-import type { IEvent } from "../utils/events";
+import type { IEvent } from "../types/events";
 import { getFormattedDate } from "../utils/events";
 import Icon from "./Icon.astro";
